test(endtoend): Fix issues with PostgreSQL tests

+++ internal/endtoend/testdata/overrides_go_types/postgresql/pgx/v5/schema.sql
@@ -2,7 +2,7 @@ CREATE TABLE foo (
     id         uuid NOT NULL,
     other_id   uuid NOT NULL,
     age        integer,
-    balance    double,
+    balance    double precision,
     bio        text,
     about      text
 );
@@ -12,7 +12,7 @@ CREATE TABLE bar (
     other_id   uuid NOT NULL,
     more_id    uuid NOT NULL,
     age        integer,
-    balance    double,
+    balance    double precision,
     bio        text,
     about      text
 );
